+++ settings/CheckoutSettings.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import _ from 'lodash';
 import { stripesShape } from '@folio/stripes-core/src/Stripes';
 import ConfigManager from '@folio/stripes-smart-components/lib/ConfigManager';
 
@@ -21,7 +22,8 @@ class CheckoutSettings extends React.Component {
   // eslint-disable-next-line class-methods-use-this
   getInitialValues(settings) {
     const value = settings.length === 0 ? '' : settings[0].value;
-    const defaultConfig = { prefPatronIdentifier: '', audioAlertsEnabled: false };
+    const defaultConfig = { prefPatronIdentifier: '', audioAlertsEnabled: false,
+                            checkoutTimeout:true, checkoutTimeoutDuration:3 };
     let config;
 
     try {
@@ -29,20 +31,24 @@ class CheckoutSettings extends React.Component {
     } catch (e) {
       config = defaultConfig;
     }
-
-    const { audioAlertsEnabled, prefPatronIdentifier } = config;
+    const { audioAlertsEnabled, prefPatronIdentifier, checkoutTimeout, checkoutTimeoutDuration } = config;
     const values = (prefPatronIdentifier) ? prefPatronIdentifier.split(',') : [];
     const idents = patronIdentifierTypes.map(i => (values.indexOf(i.key) >= 0));
 
-    return { idents, audioAlertsEnabled };
+    return { idents, audioAlertsEnabled, checkoutTimeout, checkoutTimeoutDuration };
   }
   validate(values) {
+    const stripes = this.props.stripes;
     const errors = {};
     const idents = values.idents;
     if (!idents) return errors;
     const isValid = idents.reduce((valid, v) => (valid || v), false);
     if (!isValid) {
-      errors.idents = { _error: this.props.stripes.intl.formatMessage({ id: 'ui-circulation.settings.checkout.validate.selectContinue' }) };
+      errors.idents = { _error: stripes.intl.formatMessage({ id: 'ui-circulation.settings.checkout.validate.selectContinue' }) };
+    }
+    const checkoutTimeoutDuration = (_.isInteger(+values.checkoutTimeoutDuration) && (+values.checkoutTimeoutDuration > 0))
+    if (!checkoutTimeoutDuration){
+        errors.checkoutTimeoutDuration = { _error: stripes.intl.formatMessage({ id: 'ui-circulation.settings.checkout.validate.timeoutDuration'})};
     }
     return errors;
   }