+++ CHANGELOG.md
@@ -23,8 +23,10 @@
 * Match periodId and profileId values with the values on the server. Fixes UICIRC-53.
 * Fix creating loan policy with fixed due date schedule. Fixes UICIRC-56.
 * Remove "Indefinite" loan profile. Fixes UICIRC-55.
+* Refined fixed due date schedule settings. UICIRC-40
 * Favor metadata over metaData. One case to rule them all. Refs UICIRC-57.
 * Show chosen Fixed due date schedule on loan policy details. Fixes UICIRC-58.
+* Added checkout session timeout settings in circulation->otherSettings UICIRC-48
 
 ## [1.1.1](https://github.com/folio-org/ui-circulation/tree/v1.1.1) (2017-09-02)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.1.0...v1.1.1)